requirements.txt
fastapi
uvicorn~=0.18.3
ruamel.yaml
